refactor(ui-browser): extract layout sidebar and shared entry rendering
