+++ src/ui/browser/entries.rs
@@ -1,0 +1,151 @@
+use super::super::theme::Palette;
+use super::super::{helpers, theme};
+use super::{render_grid, render_list};
+use crate::app::{App, Entry, FrameState, format_size, format_time_ago};
+use ratatui::{
+    Frame,
+    layout::Rect,
+    style::{Modifier, Style},
+    text::{Line, Span},
+    widgets::{Block, BorderType, Borders},
+};
+
+pub(super) fn render_entries(
+    frame: &mut Frame<'_>,
+    area: Rect,
+    app: &App,
+    state: &mut FrameState,
+    palette: Palette,
+) {
+    state.entries_panel = Some(area);
+    let path_text = helpers::stable_path_label(&app.cwd, area.width.saturating_sub(10) as usize);
+    let block = Block::default()
+        .borders(Borders::ALL)
+        .border_type(BorderType::Rounded)
+        .style(Style::default().bg(palette.panel_alt).fg(palette.text))
+        .border_style(Style::default().fg(palette.border));
+    frame.render_widget(&block, area);
+    helpers::render_panel_title(
+        frame,
+        area,
+        Line::from(vec![
+            Span::styled(
+                " 󰉖 ",
+                Style::default()
+                    .fg(palette.accent)
+                    .add_modifier(Modifier::BOLD),
+            ),
+            Span::styled(
+                path_text,
+                Style::default()
+                    .fg(palette.accent_text)
+                    .add_modifier(Modifier::BOLD),
+            ),
+            Span::raw(" "),
+        ]),
+    );
+    let inner = block.inner(area);
+    helpers::fill_area(frame, inner, palette.panel_alt, palette.text);
+
+    if app.view_mode == crate::app::ViewMode::Grid {
+        render_grid(frame, inner, app, state, palette);
+    } else {
+        render_list(frame, inner, app, state, palette);
+    }
+}
+
+pub(super) fn browser_entry_detail(app: &App, entry: &Entry) -> Option<String> {
+    if entry.is_dir() {
+        app.directory_item_count_label(entry)
+    } else {
+        Some(format_size(entry.size))
+    }
+}
+
+pub(super) fn browser_entry_modified(entry: &Entry) -> String {
+    entry
+        .modified
+        .map(format_time_ago)
+        .unwrap_or_else(|| "unknown".to_string())
+}
+
+pub(super) fn browser_directory_secondary(app: &App, entry: &Entry) -> String {
+    match app.directory_item_count_label(entry) {
+        Some(count) => format!("{count}  •  {}", browser_entry_modified(entry)),
+        None => browser_entry_modified(entry),
+    }
+}
+
+pub(super) fn render_compact_list_row(
+    app: &App,
+    entry: &Entry,
+    selected: bool,
+    row_width: u16,
+    palette: Palette,
+) -> Line<'static> {
+    let detail_width = 12usize;
+    let modified_width = 10usize;
+    let multi_selected = app.is_selected(&entry.path);
+    let marker_color = if selected {
+        palette.selected_border
+    } else if multi_selected {
+        palette.selection_bar
+    } else {
+        palette.panel_alt
+    };
+    let icon = theme::entry_symbol(entry);
+    let icon_style = Style::default()
+        .fg(theme::entry_color(entry, palette))
+        .add_modifier(Modifier::BOLD);
+    let name_style = if selected {
+        Style::default()
+            .fg(palette.text)
+            .add_modifier(Modifier::BOLD)
+    } else {
+        Style::default().fg(palette.text)
+    };
+    let muted_style = Style::default().fg(palette.muted);
+    let name_width = row_width
+        .saturating_sub(1 + 3 + detail_width as u16 + modified_width as u16)
+        .max(1) as usize;
+    let name = helpers::clamp_label(&entry.name, name_width);
+    let detail = pad_left(
+        browser_entry_detail(app, entry).unwrap_or_default(),
+        detail_width,
+    );
+    let modified = pad_left(browser_entry_modified(entry), modified_width);
+
+    Line::from(vec![
+        Span::styled(
+            if selected || multi_selected {
+                "▌"
+            } else {
+                " "
+            },
+            Style::default().fg(marker_color),
+        ),
+        Span::raw(" "),
+        Span::styled(icon.to_string(), icon_style),
+        Span::raw(" "),
+        Span::styled(pad_right(name, name_width), name_style),
+        Span::styled(detail, muted_style),
+        Span::styled(modified, muted_style),
+    ])
+}
+
+fn pad_left(mut text: String, width: usize) -> String {
+    let visible = helpers::display_width(&text);
+    if visible >= width {
+        return helpers::clamp_label(&text, width);
+    }
+    text = format!("{}{}", " ".repeat(width - visible), text);
+    text
+}
+
+fn pad_right(text: String, width: usize) -> String {
+    let visible = helpers::display_width(&text);
+    if visible >= width {
+        return helpers::clamp_label(&text, width);
+    }
+    format!("{text}{}", " ".repeat(width - visible))
+}
